+++ UnleashClient/features/__init__.py
@@ -1,1 +1,1 @@
-from .features import Feature
+from .Feature import Feature