+++ UnleashClient/strategies/ApplicationHostname.py
@@ -1,12 +1,12 @@
 import platform
-from UnleashClient.strategies import Strategy
+from UnleashClient.strategies.Strategy import Strategy
 
 
 class ApplicationHostname(Strategy):
     def load_provisioning(self) -> list:
         return [x.strip() for x in self.parameters["hostNames"].split(',')]
 
-    def __call__(self, context: dict = None) -> bool:
+    def apply(self, context: dict = None) -> bool:
         """
         Returns true if userId is a member of id list.