+++ UnleashClient/strategies/Default.py
@@ -1,8 +1,8 @@
-from UnleashClient.strategies import Strategy
+from UnleashClient.strategies.Strategy import Strategy
 
 
 class Default(Strategy):
-    def __call__(self, context: dict = None) -> bool:
+    def apply(self, context: dict = None) -> bool:
         """
         Return true if enabled.