+++ UnleashClient/strategies/GradualRolloutRandom.py
@@ -1,9 +1,9 @@
 import random
-from UnleashClient.strategies import Strategy
+from UnleashClient.strategies.Strategy import Strategy
 
 
 class GradualRolloutRandom(Strategy):
-    def __call__(self, context: dict = None) -> bool:
+    def apply(self, context: dict = None) -> bool:
         """
         Returns random assignment.