+++ UnleashClient/strategies/Strategy.py
@@ -1,22 +1,67 @@
 # pylint: disable=dangerous-default-value
-class Strategy():
+import warnings
+from UnleashClient.constraints import Constraint
+
+
+class Strategy:
     """
+    The parent class for default and custom strategies.
+
     In general, default & custom classes should only need to override:
-    * __call__() - Implementation of the strategy.
-    * load_provisioning - Loads strategy provisioning
+    * __init__() - Depending on the parameters your feature needs
+    * apply() - Your feature provisioning
     """
     def __init__(self,
-                 parameters: dict = {}) -> None:
+                 constraints: list = [],
+                 parameters: dict = {},
+                 ) -> None:
         """
         A generic strategy objects.
 
-        :param parameters: 'parameters' key from strategy section (...from feature section) of
+        :param constraints: List of 'constraints' objects derived from strategy section (...from feature section) of
+        /api/clients/features response
+        :param parameters: The 'parameter' objects from the strategy section (...from feature section) of
         /api/clients/features response
         """
-        # Experiment information
         self.parameters = parameters
-
+        self.constraints = constraints
+        self.parsed_constraints = self.load_constraints(constraints)
         self.parsed_provisioning = self.load_provisioning()
+
+    def __call__(self, context: dict = None):
+        warnings.warn(
+            "unleash-client-python v3.x.x requires overriding the execute() method instead of the __call__() method.",
+            DeprecationWarning
+        )
+
+    def execute(self, context: dict = None) -> bool:
+        """
+        Executes the strategies by:
+        - Checking constraints
+        - Applying the strategy
+
+        :param context: Context information
+        :return:
+        """
+        flag_state = False
+
+        if all([constraint.apply(context) for constraint in self.parsed_constraints]):
+            flag_state = self.apply(context)
+
+        return flag_state
+
+    def load_constraints(self, constraints_list: list) -> list:  #pylint: disable=R0201
+        """
+        Loads constraints from provisioning.
+
+        :return:
+        """
+        parsed_constraints_list = []
+
+        for constraint_dict in constraints_list:
+            parsed_constraints_list.append(Constraint(constraint_dict=constraint_dict))
+
+        return parsed_constraints_list
 
     # pylint: disable=no-self-use
     def load_provisioning(self) -> list:
@@ -27,14 +72,11 @@ class Strategy():
         """
         return []
 
-    def __eq__(self, other):
-        return self.parameters == other.parameters
-
-    def __call__(self, context: dict = None) -> bool:
+    def apply(self, context: dict = None) -> bool:  #pylint: disable=W0613,R0201
         """
         Strategy implementation goes here.
 
-        :param context: Context information
+        :param context:
         :return:
         """
         return False