+++ UnleashClient/strategies/__init__.py
@@ -6,3 +6,4 @@ from .GradualRolloutSessionId import GradualRolloutSessionId
 from .GradualRolloutRandom import GradualRolloutRandom
 from .RemoteAddress import RemoteAddress
 from .ApplicationHostname import ApplicationHostname
+from .FlexibleRolloutStrategy import FlexibleRollout