+++ tests/utilities/mocks/mock_custom_strategy.py
@@ -10,7 +10,49 @@ MOCK_CUSTOM_STRATEGY = {
           "name": "amIACat",
           "parameters": {
             "sound": "meow,nyaa"
+          },
+          "constraints": [
+            {
+              "contextName": "environment",
+              "operator": "IN",
+              "values": [
+                "staging",
+                "prod"
+              ]
+            }
+          ]
+        }
+      ],
+      "createdAt": "2018-10-13T10:15:29.009Z"
+    },
+    {
+      "name": "CustomToggleWarning",
+      "description": "CustomToggle Warning Test",
+      "enabled": True,
+      "strategies": [
+        {
+          "name": "amIADog",
+          "parameters": {
+            "sound": "arf,bark"
           }
+        }
+      ],
+      "createdAt": "2018-10-13T10:15:29.009Z"
+    },
+    {
+      "name": "CustomToggleWarningMultiStrat",
+      "description": "CustomToggle Warning Test",
+      "enabled": True,
+      "strategies": [
+        {
+          "name": "amIADog",
+          "parameters": {
+            "sound": "arf,bark"
+          }
+        },
+        {
+          "name": "default",
+          "parameters": {}
         }
       ],
       "createdAt": "2018-10-13T10:15:29.009Z"